修改：     core/arms.hpp 	修改：     core/ensemble_sample.hpp 	修改：     core/forward_sampleable_node.hpp 	修改：     core/gibbs_sampler.hpp 	修改：     core/mcmc_traits.hpp 	修改：     core/particle_filter.hpp 	修改：     core/particle_smoother.hpp 	修改：     core/ptransform.hpp 	修改：     core/ptsample.hpp 	修改：     test_suit/test_bimodal.cpp

+++ core/forward_sampleable_node.hpp
@@ -88,7 +88,7 @@ namespace mcmc_utilities
 	    {
 	      return result;
 	    }
-	  for(int i=0;i<p->num_of_dims();++i)
+	  for(size_t i=0;i<p->num_of_dims();++i)
 	    {
 	      result=result||p->is_observed(i);
 	      if(result)
